requirements.txt
sphinx==7.2.6
furo
myst-parser
python-dotenv
dvc[s3]
roboflow
boto3
mlflow
ultralytics
requests
onnxruntime==1.20.0